requirements.txt
asgiref==3.4.1
DateTime==4.3
Django==3.2.5
djangorestframework==3.12.4
ImageHash==4.2.1
numpy==1.21.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pytz==2021.1
PyWavelets==1.1.1
scipy==1.7.0
simhash==2.0.0
six==1.16.0
sqlparse==0.4.1
zope.interface==5.4.0
pip~=21.1.3